fix(app): pr review
