+++ app-near/src/server/users/connection.ts
@@ -30,7 +30,7 @@ export const login = async (
   const userSurvey = await db.survey.findFirst();
 
   if (!userSurvey) {
-    console.debug(`Aucune enquête trouvée pour ${email}, accès non autorisé.`);
+    console.debug(`Aucune enquête trouvée accès non autorisé.`);
     return { message: "Accès non autorisé", success: false };
   }
