+++ app-near/src/types/enums/broadcasting.ts
@@ -1,0 +1,7 @@
+export type SurveyType = "su" | "ngc" | "mdv" | "mdd";
+
+export type BroadcastType =
+  | "mail_campaign"
+  | "social_network"
+  | "street_survey"
+  | "qr_code";
